+++ tap_circle_ci/streams/deploy.py
@@ -1,7 +1,7 @@
 from typing import Dict, Iterator
 from datetime import datetime
 from singer import Transformer, get_logger, metrics, write_record
-from .abstracts import IncrementalStream
+from .abstracts import IncrementalStream, _iter_pages
 
 LOGGER = get_logger()
 
@@ -31,27 +31,14 @@ class Deploy(IncrementalStream):
         return self.url_endpoint.format(organization_id=organization_id, page_size=page_size)
 
     def get_records(self) -> Iterator[Dict]:
+        """Fetch all records for each organization with pagination."""
         org_ids = self.get_org_ids()
         with metrics.Counter("page_count") as counter:
             for org_id in org_ids:
                 url = self.get_url(org_id)
-                # ---- First page ----
-                params = {}  # reset params for every org
-                response = self.client.get(url, params, {})
-                counter.increment()
-                items = response.get("items", [])
-                next_page_token = response.get("next_page_token")
-                for record in items:
-                    record["organization_id"] = org_id
-                    yield record
-                # ---- Subsequent pages ----
-                while next_page_token:
-                    params = {"page-token": next_page_token}  # new dict each loop
-                    response = self.client.get(url, params, {})
+                for page in _iter_pages(self.client.get, url):
                     counter.increment()
-                    items = response.get("items", [])
-                    next_page_token = response.get("next_page_token")
-                    for record in items:
+                    for record in page.get("items", []):
                         record["organization_id"] = org_id
                         yield record
