+++ tap_circle_ci/streams/project.py
@@ -1,6 +1,6 @@
 from typing import Dict, Iterator, List
 from singer import metrics, write_record, get_logger
-from .abstracts import FullTableStream
+from .abstracts import FullTableStream, _iter_pages
 
 LOGGER = get_logger()
 
@@ -29,26 +29,14 @@ class Project(FullTableStream):
     def get_records(self) -> Iterator[Dict]:
         """Fetch all project records for each org (supports pagination)."""
         org_ids = self.get_org_ids()
+
         with metrics.Counter("page_count") as counter:
             for org_id in org_ids:
                 url = self.get_url(org_id)
-                # ---- First page ----
-                params = {}
-                response = self.client.get(url, params, {})
-                counter.increment()
-                items = response.get("items", [])
-                next_page_token = response.get("next_page_token")
-
-                for item in items:
-                    item["organization_id"] = org_id
-                    yield item
-                # ---- Subsequent pages ----
-                while next_page_token:
-                    params = {"page-token": next_page_token}
-                    response = self.client.get(url, params, {})
+                # Loop through all pages using the helper
+                for page in _iter_pages(self.client.get, url):
                     counter.increment()
-                    items = response.get("items", [])
-                    next_page_token = response.get("next_page_token")
+                    items = page.get("items", [])
                     for item in items:
                         item["organization_id"] = org_id
                         yield item
@@ -60,7 +48,10 @@ class Project(FullTableStream):
                 for record in self.get_records():
                     transformed = transformer.transform(record, schema, stream_metadata)
                     write_record(self.tap_stream_id, transformed)
-                    all_records.append({"id": record["id"], "slug": record["slug"]})
+                    all_records.append({
+                        "id": record.get("id"),
+                        "slug": record.get("slug")
+                    })
                     counter.increment()
 
         # Store both id and slug for other streams to use
